# improve object/jstring parser