@phadej 2016-07-15
I'm planning to do old-aeson/new-aeson change to `Typed` benchmark, as in https://github.com/bos/aeson/pull/447. So we can see the difference better.