@winterland1989 2016-07-15
> I'm planning to do old-aeson/new-aeson change to Typed benchmark, as in #447. So we can see the difference better.

@phadej  Can you tell me what changes should i make?