@winterland1989 2016-07-18
I'd like to drop`attoparsec` in favor of non-backtrace parsers to further optimize parsing speed, attoparse is really not needed here, but that will break existing code using `Data.Aeson.Parser`, we need a discuss here.