@phadej 2016-07-18
Are you sure it will improve a lot? Even the `attoparsec` is backtracking, we don't backtrack in aeson. Yet benchmark result would prove/disproof this