@ondrap 2016-07-18
I think the author of `scanner` mentioned somewhere that he has a json parser and that it is ultimately faster.