@winterland1989 2016-07-18
Yes, `scanner` is exactly what i'm using, stay turned, benchmark comming!