@winterland1989 2016-07-20
I don't have enough time to finish the non backtracking experiment quickly, so i leave the code as it now.