@bergmark 2016-08-07
@winterland1989 let's continue here instead of #392. I was referring to your comment

> I will post the benchmark figures before further commits.

Which benchmarks did you mean?